Quantum_chat_app/requirements.txt
streamlit
fastapi
uvicorn
websockets
pymongo
motor